Fixed minus sign after 0 justification and intmax issues
The zeros were being put before the minus sign of the number. That was fixed by adding an if case to set the minus first and remove it from the number to be written later.
INTMAX and MIN were having problem as get number was working with int. It was set to work with ssize_t to cover all max and min and argument being sent was casted to that

+++ ft_printf_bonus.h
@@ -18,6 +18,6 @@ int		print_space(va_list *l, const char *s, int *i);
 int		print_signal(va_list *l, const char *s, int *i);
 void	get_unsig(unsigned int n, char base, char *conv);
 void	get_unsig_long(unsigned long n, char *conv);
-void	get_nbr(int n, char *conv, int sign);
+void	get_nbr(ssize_t n, char *conv, int sign);
 
 #endif
